refactor: move components test files
